Remove unneeded `is rw`
It used to work with `is rw` in previous Rakudo releases even
though it was incorrect. Future rakudo releases may include a fix, so
changing it according to that. Basically, the Hash object there is
already writable so `is rw` is not needed. Without `is rw` it works on
all v6.c rakudo releases.